Break up tokenized code writer
