+++ src/PhpSpec/CodeGenerator/Generator/MethodGenerator.php
@@ -15,10 +15,10 @@ namespace PhpSpec\CodeGenerator\Generator;
 
 use PhpSpec\Console\IO;
 use PhpSpec\CodeGenerator\TemplateRenderer;
-use PhpSpec\Util\CodeWriter;
+use PhpSpec\CodeGenerator\Writer\CodeWriter;
 use PhpSpec\Util\Filesystem;
 use PhpSpec\Locator\ResourceInterface;
-use PhpSpec\Util\TokenizedCodeWriter;
+use PhpSpec\CodeGenerator\Writer\TokenizedCodeWriter;
 
 /**
  * Generates class methods from a resource
