+++ src/PhpSpec/CodeGenerator/Writer/CodeWriter.php
@@ -11,7 +11,7 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpSpec\Util;
+namespace PhpSpec\CodeGenerator\Writer;
 
 interface CodeWriter
 {
